backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
python-dotenv==1.0.1
pypdf==4.3.1
requests==2.32.3
numpy==2.0.1
scikit-learn==1.5.1
sentence-transformers==3.0.1

